V1.2 adding comments, fixing bugs, isolating decode
